Publish Minion by Brian Handal
This is my first push of the minion gui. here are some included features
with this repo:
-user authentication; currently set to personalized network, which needs
to be altered in order to work with other group
-SQL integration to store server and analysis history information. User
will need to setup their own SQL database in order to get this feature
to work. Without database the analysis will not run due to the fact that
it looks to save run information in the database.
-SharpSSH, a reference library provided by Tamir Gal to facilitate the
ssh process. This library is used to test server existence and send run
commands without interfacing with the server directly
-NextGen Bioinformatic tools. The great advantage of Minion is the fact
that it allows the user to select any pipeline to perform data analysis.
Although  bwa and samtools is the only pipeline that has been optimized,
more pipelines will be provided soon.

Summary of action: Minion is a windows based graphical interface that
engages the user to perform ssh protocols and perform bioinformatic
procedures without having to directly execute command line arguments. A
user will simply login, connect to the desired server, and the runs on
that server will be parsed into a combobox for selection. Once a run is
selected a form prompts the user to select their desired pipeline
settings and the proceed with the analysis. Minion's core function is
handled by a backend python script that receives the user input and
performs the bioinformatic procedures based upon the settings given.

Soon to come
-more analysis pipelines other than just bwa and samtools, as well as
other qc features such as fastqc and bamstats.
-NoSQL based viewing port to visualize analyzed runs. This is simply a
feature that parses vcf files in to a table format to be easily viewed.
Also, the user will be able to select a variant and it will redirect to
the integrated genome viewer for subsequent interpretation.