Minion v1.0
This update includes the new Minion interface that enables users to
create a custom run across fastq files that are not necessarily in the
same run.

Items of concern:
-How to connect to a server to analyze this run. Since this is done via
a custom tabe there is no need to connect to a server, so I may create
another ssh tab that allows the users to connect to a server. This may
be best. Once you are connected it will enable the Main and custom tab